+++ apps/fluent-tester/src/FluentTester/TestComponents/Avatar/CustomizedAvatar.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
-import { JSAvatar, IconAlignment } from '@fluentui-react-native/experimental-avatar';
+import { JSAvatar, IconAlignment, AvatarSize } from '@fluentui-react-native/experimental-avatar';
 import { Switch, View, Text, TextInput } from 'react-native';
-import { Slider } from '../Common/Slider';
 import { steveBallmerPhotoUrl } from './../PersonaCoin/styles';
 import { AlignmentPicker } from '../Common/AlignmentPicker';
 import { commonTestStyles as commonStyles } from '../Common/styles';
@@ -10,13 +9,13 @@ export const CustomizeUsage: React.FunctionComponent = () => {
   const [showImage, setShowImage] = React.useState(true);
   const [coinColor, setCoinColor] = React.useState<string>();
   const [textColor, setTextColor] = React.useState<string>();
-  const [size, setSize] = React.useState<number>(96);
+  const [size, setSize] = React.useState<AvatarSize>(96);
   const [iconSize, setIconSize] = React.useState<number>(24);
   const [initialsSize, setInitialsSize] = React.useState<number>(14);
   const [horizontalAlignment, setHorizontalAlignment] = React.useState<IconAlignment>();
   const [verticalAlignment, setVerticalAlignment] = React.useState<IconAlignment>();
 
-  const [borderColor, setBorderColor] = React.useState<string>('red');
+  const [ringColor, setRingColor] = React.useState<string>('red');
   const [ringBackgroundColor, setRingBackgroundColor] = React.useState<string>(undefined);
   const [showRing, setShowRing] = React.useState<boolean>(true);
   const [transparent, setTransparent] = React.useState<boolean>(false);
@@ -29,12 +28,12 @@ export const CustomizeUsage: React.FunctionComponent = () => {
       verticalIconAlignment: verticalAlignment,
       iconSize: iconSize,
       initialsSize: initialsSize,
-      width: size,
-      height: size,
-      borderColor,
+      size,
+      ringColor,
+      ringBackgroundColor,
     };
     return JSAvatar.customize(tokens);
-  }, [coinColor, textColor, horizontalAlignment, verticalAlignment, iconSize, initialsSize, size, borderColor]);
+  }, [coinColor, textColor, horizontalAlignment, verticalAlignment, iconSize, initialsSize, size, ringColor, ringBackgroundColor]);
 
   return (
     <View style={commonStyles.root}>
@@ -76,7 +75,35 @@ export const CustomizeUsage: React.FunctionComponent = () => {
           placeholder="Ring color"
           blurOnSubmit={true}
           onSubmitEditing={(e) => {
-            setBorderColor(e.nativeEvent.text);
+            setRingColor(e.nativeEvent.text);
+          }}
+        />
+
+        <TextInput
+          style={[commonStyles.textBox]}
+          placeholder="Avatar size"
+          blurOnSubmit={true}
+          onSubmitEditing={(e) => {
+            const size = e.nativeEvent.text as unknown as AvatarSize;
+            setSize(size);
+          }}
+        />
+
+        <TextInput
+          style={[commonStyles.textBox]}
+          placeholder="Icon size"
+          blurOnSubmit={true}
+          onSubmitEditing={(e) => {
+            setIconSize(parseInt(e.nativeEvent.text));
+          }}
+        />
+
+        <TextInput
+          style={[commonStyles.textBox]}
+          placeholder="Initials size"
+          blurOnSubmit={true}
+          onSubmitEditing={(e) => {
+            setInitialsSize(parseInt(e.nativeEvent.text));
           }}
         />
 
@@ -89,17 +116,9 @@ export const CustomizeUsage: React.FunctionComponent = () => {
           }}
         />
 
+        <Text>These features will be available later</Text>
         <AlignmentPicker style={commonStyles.header} label="Horizontal icon alignment" onSelectionChange={setHorizontalAlignment} />
         <AlignmentPicker style={commonStyles.header} label="Vertical icon alignment" onSelectionChange={setVerticalAlignment} />
-
-        <Text>Coin size</Text>
-        <Slider minimum={8} maximum={200} initialValue={80} style={commonStyles.vmargin} onChange={setSize} />
-
-        <Text>Icon size</Text>
-        <Slider minimum={8} maximum={100} initialValue={24} style={commonStyles.vmargin} onChange={setIconSize} />
-
-        <Text>Font size</Text>
-        <Slider minimum={5} maximum={50} initialValue={14} style={commonStyles.vmargin} onChange={setInitialsSize} />
       </View>
 
       <CustomizedAvatar
@@ -109,11 +128,9 @@ export const CustomizeUsage: React.FunctionComponent = () => {
         accessibilityLabel="Former CEO of Microsoft"
         badge={{ status: 'blocked' }}
         src={showImage ? steveBallmerPhotoUrl : undefined}
-        size={56}
         ring={
           showRing
             ? {
-                ringBackgroundColor,
                 ringThickness: 4,
                 innerGap: 4,
                 transparent,