+++ packages/theming/theming-utils/src/index.ts
@@ -1,2 +1,3 @@
 export { getCurrentAppearance } from './getCurrentAppearance';
 export { mapPipelineToTheme, mapFontPipelineToTheme } from './mapPipelineToTheme';
+export { mapPipelineToShadow } from './mapPipelineToShadow';