+++ packages/theming/default-theme/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Change Log - @fluentui-react-native/default-theme
 
-This log was last generated on Thu, 28 Apr 2022 19:09:51 GMT and should not be manually modified.
+This log was last generated on Sat, 21 May 2022 01:47:57 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.0
+
+Sat, 21 May 2022 01:47:57 GMT
+
+### Minor changes
+
+- Add shadows to default (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/theme-tokens to v0.16.3
+- Bump @fluentui-react-native/theme-types to v0.15.0
+- Bump @fluentui-react-native/theme to v0.6.9
+- Bump @fluentui-react-native/theming-utils to v0.11.0
 
 ## 0.10.3